requirements.txt
flask
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
selenium
playwright
pyppeteer
apscheduler
sendgrid
slackclient
pytest
